requirements.txt
alabaster==0.7.12
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
azure-common==1.1.24
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
Babel==2.8.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
dj-pagination==2.4.0
Django==2.2
django-allauth==0.40.0
django-annoying==0.10.5
django-cors-middleware==1.4.0
django-crispy-forms==1.8.0
django-debug-toolbar==2.0
django-elasticsearch-dsl==0.5.1
django-extensions==2.2.5
django-filter==2.2.0
django-formtools==2.1
django-gravatar2==1.4.2
django-heroku==0.3.1
django-kombu==0.9.4
django-messages-extends==0.6.0
django-polymorphic==2.1.2
django-storages==1.7.2
django-taggit==1.1.0
django-textclassifier==1.0
django-vanilla-views==1.0.6
djangorestframework==3.10.3
djangorestframework-jsonp==1.0.2
dnspython==1.16.0
docker==4.1.0
docutils==0.16
drf-extensions==0.5.0
drf-flex-fields==0.6.1
elasticsearch==6.4.0
elasticsearch-dsl==6.1.0
et-xmlfile==1.0.1
gitdb2==2.0.6
GitPython==3.0.4
gunicorn==20.0.4
heroku==0.1.4
idna==2.8
imagesize==1.2.0
importlib-metadata==1.4.0
jdcal==1.4.1
Jinja2==2.11.0
jsonfield==2.0.2
kombu==4.6.7
MarkupSafe==1.1.1
mock==3.0.5
mysql-connector-python==8.0.19
nltk==3.4.5
oauthlib==3.1.0
openpyxl==3.0.2
orjson==2.0.7
packaging==19.2
Pillow==6.2.1
protobuf==3.6.1
psycopg2==2.7.5
psycopg2-binary==2.8.4
pycparser==2.19
pycrypto==2.6.1
Pygments==2.4.2
pyparsing==2.4.6
python-dateutil==1.5
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
redis==3.3.11
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
selectolax==0.2.1
six==1.14.0
slumber==0.7.1
smmap2==2.0.5
snowballstemmer==2.0.0
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.2
sqlparse==0.3.0
stripe==2.38.0
textblob==0.15.3
ua-parser==0.8.0
Unidecode==1.1.1
Unipath==1.1
urllib3==1.25.3
user-agents==2.0
vine==1.3.0
virtualenv==16.7.7
websocket-client==0.57.0
whitenoise==5.0.1
xlrd==1.2.0
XlsxWriter==1.2.1
zipp==2.1.0
